Improve install guide navigation
Current the Install guide index page uses a form with radios to route people to one of three pages.

This replaces the form with links, as those are generally better for navigation (for example can be opened in a new tab).

It also combines the "Download the kit" page with the advanced install guide to simplify things.