Switch to using an includes instead of custom macro
I think a Nunjucks `includes` is a simpler option here, as there are no params being set in the macro.

Also switched to use the `insetText` macro instead of HTML, and reduced the padding and margins a bit.

Tweaked the wording in the inset text to remove the repetition of "NHS England Windows laptop".