Rename 'How to guides' as 'Guides'
This renames the 'How to guides' as simply 'Guides'.

This is because:

* 'How to guides' is an awkward phrase to read, as you. might initially assume the third word is a verb (like "How to code" and then have to re-evaluate it as a noun phrase
* It’s longer than it needs to be, and we might want to have a navigation bar at some point, in which case 'Guides' would be a lot shorter

In addition, some of the guides themselves (which had a mixture of forms) have been renamed to no longer include 'How to'. This also means the link text on the index page now matches the titles when you click through.

They all now start with a verb (although a mix of tenses).

The changes are:

* How to setup Git -> Setting up Git
* How to use components -> Using components
* How to do branching -> Branching
* How to make pages -> Making pages
* How to pass data page to page -> Passing data page to page

The other guide titles are unchanged:

* Build a basic prototype using the NHS.UK prototype kit
* Adding CSS, JavaScript and Images
* Updating the kit
* Switching from the GOV.UK prototype kit

The breadcrumb is also changed from "How to guides" to just "Guides"