First pass
This brings the guidance into the prototype kit and also removes the examples as we tend not to encourage them anymore.

To consider - is there still duplication (for example do we need the session data at all)? can this be split into multiple sections (I'm not sure as I feel they overlap so much)