Add guide for using GitHub Codespaces (#152)
This PR is to add instructions for using the NHS Prototype Kit with
Github Codespaces which is intended as a viable option for people who
cannot (or would prefer not to) install the prototype kit on their
computers.

It is contingent on [a separate
PR](https://github.com/nhsuk/nhsuk-prototype-kit/pull/428) being merged
to the prototype kit for a devcontainer to enable this.

## Get started page


![image](https://github.com/user-attachments/assets/5a573d78-b8a3-4011-825d-aa9a5ae7ca12)

## Setup the prototype kit on GitHub Codespaces


![image](https://github.com/user-attachments/assets/f5f5b707-f574-452c-84bb-9dbb82a2be33)

## Step 1 - Create a Codespace


![image](https://github.com/user-attachments/assets/db3a4d4c-c295-44a6-af45-6ea7620e0c20)

## Step 2 - Working with Codespaces


![image](https://github.com/user-attachments/assets/6287fc1a-d4b5-46ed-a45d-b5b0c2e0786a)

---------

Co-authored-by: vickytnz <vicky.teinaki@gmail.com>
Co-authored-by: Sara Wilcox <45792959+sarawilcox@users.noreply.github.com>
Co-authored-by: anandamaryon1 <anandamaryon@gmail.com>
Co-authored-by: Frankie Roberto <frankie@frankieroberto.com>