Include auto-store-data script
This script adds a hidden checkbox to all forma containing checkboxes, with the same name but a value of `_unchecked`.

This is useful for making sure that the values are always stored in an array, even when a single item is checked, as well as being able to more easily update data when a checkbox is unchecked.