Update guide and tutorial to use new values option
Version [9.2.0](https://github.com/nhsuk/nhsuk-frontend/releases/tag/v9.2.0) of NHS Frontend includes a simpler way to set the pre-selected or pre-checked options for checkboxes and radios.

This let you set `value: data.attribute` instead of having to use the `checked()` filter on each item within a checkbox or radio.

The guide and examples can now be updated to use this.