Update guides to use new 'value' option instead of 'checked' filter (#223)
Since [version 9.2.0 of NHS
Frontend](https://github.com/nhsuk/nhsuk-frontend/releases/tag/v9.2.0)
there’s now an easier, nicer way of setting the default selected radio
option or default checked checkboxes.

Instead of using the `checked()` filter on each `item` within radios and
checkboxes, the `value` (for radios) or `values` (for checkbox) of the
selected items can be set at the top level of the Nunjucks params.

This pull request updates the guidance to use the newer, easier way of
doing this.

The `checked` function/filter is also deleted from the guidance site as
it’s no longer needed (but will remain in the kit itself for
backwards-compatibility until at least the next major release).

---------

Co-authored-by: Colin Rotherham <work@colinr.com>