Simplify guides index page
This has a lot on it right now, including links to all the Getting started guides (which are in another section of the site), and each step of the multi-part tutorial and using Git guides.

To make this a bit more browsable, this reduces the links to only the guides in this section, and only the first page of each of the multi-part guides.

The tutorial is also given a slightly different treatment (using a 'card') to give it a bit more prominence.