Simplify guides index page (#239)
This has a lot on it right now, including links to all the Getting
started guides (which are in another section of the site), and each step
of the multi-part tutorial and using Git guides.

To make this a bit more browsable, this reduces the links to only the
guides in this section, and only the first page of each of the
multi-part guides.

The tutorial is also given a slightly different treatment (using a
'card') to give it a bit more prominence.

The intro sentence has also been dropped.

## Screenshots

| Before | After |
| -------|-------|
| ![prototype-kit service-manual nhs
uk_how-tos](https://github.com/user-attachments/assets/ab216bf5-09b5-4b93-a7c9-ba248c5a5c4e)
| ![localhost_3000_how-tos
(1)](https://github.com/user-attachments/assets/c7f66719-7fa1-466a-b6d7-3c03ffed186e)
|