Update Get started page and initial flow
This changes the Get started page to now use 4 cards for the 4 different install guides:

* Mac
* Windows
* Coding in a browser (Codespaces)
* Advanced

This hopefully makes the Codespaces guide more discoverable, now that we’re more confidence that it’s a viable option for users.

The updated page also removes the radio branching question to get to the Mac on Windows guides, as these can be directly linked to now instead.