Add a guide for creating a new repository
This adds a guide to for creating a new repository using the GitHub website template repository feature.

We’ve found this to be a lot simpler and easier than explaining how to download the kit and then later sync it with a new empty repository on GitHub.

It also avoids a common mistake people make which is to clone the repository instead, and then get an error when trying to sync it as they do not have permissions to push.